Only remove custom actions from the actions table and added a confirmation modal for the deletion of actions

+++ app/Http/Controllers/PipelineController.php
@@ -78,8 +78,10 @@ class PipelineController extends Controller
         // Detach from the pivot.
         $project->actions()->detach($action->id);
 
-        // Delete the action.
-        $action->delete();
+        // Only delete the action if it is a custom action.
+        if ($action->custom) {
+            $action->delete();
+        }
 
         return response(null, Response::HTTP_OK);
     }
